+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Expose the farmOS instance name, version, and system of measurement
+- Add support for simple maps using farmOS-map
 
 ### Fixed